Fix computing transitive dependencies for consumed resources
Before created resources would properly be recursive as they would
call use to emit to the environment for related resources. However
this is not the case for consumed resources, meaning that we could
miss related resources for related resources.

Further since this was refactoring the code has been moved to env as
the operations are principally about the environment